Moved `utils.js` to `utils/utils.ts`
