+++ src/bandcamp/pages/page-music.js
@@ -1,5 +1,5 @@
 import { contentChangeWithPolling, createDatalistFromArray, createElementFromHTML, input, selectElementWithContent, setDataAttribute } from "../../modules/html.js";
-import { containsOneOf, splitString, isEmptyArray, countOccurrences, removeBrackets } from "../../modules/utils.js";
+import { containsOneOf, splitString, isEmptyArray, countOccurrences, removeBrackets } from "../../utils/utils";
 import { getBandPhotoSrc, getReleasesData } from "../modules/html.js";
 
 // Setup logic for BC music page
