+++ src/discogs/modules/initialization.js
@@ -1,7 +1,7 @@
 'use strict';
 
 import { click } from "../../modules/html.js";
-import { convertNewlinesToBreaks } from "../../modules/utils.js";
+import { convertNewlinesToBreaks } from "../../utils/utils";
 import { setSectionHint, fillDurations, getSubmissionFormSectionNotes, selectFormatDescription, selectFormatFileType, setInputValue, getArtistNameInput, getQuantityInput, getNotesTextarea, getSubmissionNotesTextarea } from "./draft-page.js";
 import { showNotificationError, showNotificationInfo } from "./notification.js";
