+++ src/modules/storage.js
@@ -1,7 +1,7 @@
 import { Release } from "../app/release.js";
 import { isValidBandcampURL } from "../bandcamp/modules/html.js";
 import { generateKeyForRelease, generateKeyForUrl, generateKeyUrlMapFromUrls, generateKeysFromUrls } from "./key-generator.js";
-import { isArray, isFunction, isObject } from "./utils.js";
+import { isArray, isFunction, isObject } from "../utils/utils";
 
 const storage = chrome.storage.local;
