+++ src/popup/helpers.js
@@ -4,7 +4,7 @@ import { getSearchDiscogsReleaseUrl } from "../discogs/modules/discogs.js";
 import { chromeSendMessageToCurrentTab } from "../utils/chrome";
 import { createIconLink, disable, enable, getDataAttribute, hasDataAttribute, setDataAttribute } from "../modules/html.js";
 import { generateKeyForReleaseItem } from "../modules/key-generator.js";
-import { convertToAlias, isArray, isObject, isString } from "../modules/utils.js";
+import { convertToAlias, isArray, isObject, isString } from "../utils/utils";
 
 export function createBootstrapCheckbox(id, value, labelText, checked) {
   // Create the checkbox input element
